Slash Commands

+++ src/plugins/commander/commands/javascript.js
@@ -1,9 +1,13 @@
+const { SlashCommandBuilder } = require('@discordjs/builders');
 const Command = require('../structs/Command.js');
+
+const JS_ID = '269869828691591169';
 
 class JavaScriptCommand extends Command {
     constructor(bot) {
         super(bot);
         this.aliases = ['javascript', 'js', 'rmjavascript', 'rmjs'];
+        this.schema = new SlashCommandBuilder();
 
         this.shortdesc = `Gives you the JavaScript role.`;
         this.desc = `
@@ -14,14 +18,25 @@ class JavaScriptCommand extends Command {
         ];
     }
 
-    async call(message, content) {
+    async call(message, content, { interaction }) {
         if (content) return;
 
-        await message.delete();
-        if (message.member.roles.cache.has('269869828691591169')) {
-            return message.member.roles.remove('269869828691591169');
+        const had = message.member.roles.cache.has(JS_ID);
+        if (had) {
+            await message.member.roles.remove(JS_ID);
         } else {
-            return message.member.roles.add('269869828691591169');
+            await message.member.roles.add(JS_ID);
+        }
+
+        if (interaction) {
+            await interaction.reply({
+                content: had
+                    ? `The <@&${JS_ID}> role has been taken away`
+                    : `You have been given the <@&${JS_ID}> role`,
+                ephemeral: true
+            });
+        } else {
+            await message.delete();
         }
     }
 }
