+++ src/plugins/commander/commands/portability.js
@@ -1,9 +1,13 @@
+const { SlashCommandBuilder } = require('@discordjs/builders');
 const Command = require('../structs/Command.js');
+
+const PORTABILITY_ID = '311612168061714432';
 
 class PortabilityCommand extends Command {
     constructor(bot) {
         super(bot);
         this.aliases = ['portability', 'rmportability'];
+        this.schema = new SlashCommandBuilder();
 
         this.shortdesc = `Gives you the Portability role.`;
         this.desc = `
@@ -14,14 +18,25 @@ class PortabilityCommand extends Command {
         ];
     }
 
-    async call(message, content) {
+    async call(message, content, { interaction }) {
         if (content) return;
 
-        await message.delete();
-        if (message.member.roles.cache.has('311612168061714432')) {
-            return message.member.roles.remove('311612168061714432');
+        const had = message.member.roles.cache.has(PORTABILITY_ID);
+        if (had) {
+            await message.member.roles.remove(PORTABILITY_ID);
         } else {
-            return message.member.roles.add('311612168061714432');
+            await message.member.roles.add(PORTABILITY_ID);
+        }
+
+        if (interaction) {
+            await interaction.reply({
+                content: had
+                    ? `The <@&${PORTABILITY_ID}> role has been taken away`
+                    : `You have been given the <@&${PORTABILITY_ID}> role`,
+                ephemeral: true
+            });
+        } else {
+            await message.delete();
         }
     }
 }
